wpa_supplicant — 'ifdown' inhibits ability to SCAN or REASSOCIATE
Shortly after wpa_supplicant v0.7.2, changes were introduced to support
rf_kill.

A side of effect of these changes is that wi-fi radio operations will
not be permitted unless the network interface is in the UP state.
wpa_cli will return FAIL if any attempt is made to SCAN or REASSOCIATE,
and it is likely DISCONNECT may not be acted upon, either.

The SqueezePlay/Squeeze-OS network management processes require that
SCAN and REASSOCIATE commands operate when the network interface is
*not* UP.

This change adds a network/if-post-down action script that brings the
network back up following an 'ifdown' command.

At present this is only applicable to the Squeezebox Radio (baby). The
relevant interface is eth1, this is hard coded into the rest of the
firmware.

The script makes use of the 'ip' command in preference to ‘ifconfig' as
it can be used to clear an assigned ip address. 'ip' is available in
the Squeeze-OS firmware.

The Squeezebox Controller (jive) and Touch (fab4) use different wi-fi
interfaces, so the script would require suitable amendment if ever
adopted for those platforms.

Source commits in hostapd:

Add Linux rfkill support - Commit #8401a6
https://w1.fi/cgit/hostap/commit/?id=8401a6b028c464a427654496b7b4a4cfb4ab8492

rfkill: Use rtnetlink ifup/ifdown events - Commit #a63063
https://w1.fi/cgit/hostap/commit/?id=a63063b4f6a06d110e7168560131ea52882f0474


This change is similar in operation to the ifupdown support function
'if_post_down_up' provided with the debian wpa_supplicant package:

## keep IFACE scanning
# After ifdown, the IFACE may be left "down", and inhibits 
# wpa_supplicant's ability to continue roaming.
#
# NB: use iproute if present, flushing the IFACE first
#
if_post_down_up () {
	if [ -x /bin/ip ]; then
		ip addr flush dev "$WPA_IFACE" 2>/dev/null
		ip link set "$WPA_IFACE" up
	else
		ifconfig "$WPA_IFACE" up
	fi
}